+++ src/rcpp_reincarnate.cpp
@@ -7,7 +7,7 @@
 //' @param fishpop,fishpop_track Matrix with fishpop and starting fishpop values.
 //' @param seafloor Matrix with seafloor values.
 //' @param fish_id,cell_id Vector with id of fish and corresponding cell ids.
-//' @param pop_linf,pop_n_body,pop_want_reserves Numeric with parameters.
+//' @param pop_linf,pop_n_body,pop_max_reserves Numeric with parameters.
 //' @param reason String with reason of reincarnation.
 //'
 //' @details
@@ -23,7 +23,7 @@
 void rcpp_reincarnate(Rcpp::NumericMatrix fishpop, Rcpp::NumericMatrix fishpop_track,
                       Rcpp::NumericMatrix seafloor,
                       int fish_id, int cell_id,
-                      double pop_linf, double pop_n_body, double pop_want_reserves,
+                      double pop_linf, double pop_n_body, double pop_max_reserves,
                       String reason) {
 
   // save current original coordinates
@@ -53,7 +53,7 @@ void rcpp_reincarnate(Rcpp::NumericMatrix fishpop, Rcpp::NumericMatrix fishpop_t
   fishpop(fish_id, 3) = y_coord;
 
   // calculate new reserves for new fish of new size
-  double reserves_wanted = pop_n_body * fishpop(fish_id, 6) * pop_want_reserves;
+  double reserves_wanted = pop_n_body * fishpop(fish_id, 6) * pop_max_reserves;
 
   // detritus pool is smaller than wanted reserves, detritus pool is fully used
   if (reserves_wanted >= seafloor(cell_id, 5)) {
@@ -94,12 +94,3 @@ void rcpp_reincarnate(Rcpp::NumericMatrix fishpop, Rcpp::NumericMatrix fishpop_t
 
   }
 }
-
-/*** R
-rcpp_reincarnate(NumericMatrix fishpop,
-                 Rcpp::NumericMatrix fishpop_track,
-                 Rcpp::NumericMatrix seafloor,
-                 int fish_id, int cell_id,
-                 double pop_linf, double pop_n_body, double pop_want_reserves,
-                 String reason)
-*/