+++ NEWS.md
@@ -14,6 +14,7 @@
 * Calling all `rcpp_` functions directly if possible in `run_simulation`
 * Renamed `detritus_ratio` to `seagrass_slough`
 * Renamed `detritus_fish_ratio` to `detritus_fish_decomp`
+* Remove `pop_want_reserves` parameter. Fish always try to fill up reserves
 
 # arrR 0.4.0
 * Add `simulate_input` and `simulate_output`